requirements.txt
# First install this Python version must  ( 3.9.19)
streamlit
python-dotenv
PyPDF2
langchain
faiss-cpu  
langchain-huggingface
